Restructured project, added ConventionExtension attempt at solving gradle test kit coverage issue, nope
