+++ .github/workflows/pre-merge.yaml
@@ -9,7 +9,7 @@ on:
       - '*'
 
 jobs:
-  gradle:
+  check-plugin:
     runs-on: ubuntu-latest
     if: ${{ !contains(github.event.head_commit.message, 'ci skip') }}
     steps:
@@ -20,4 +20,4 @@ jobs:
         uses: gradle/gradle-build-action@v2
 
       - name: Run Gradle tasks
-        run: ./gradlew preMerge --continue
+        run: ./gradlew test-projects:kover-badge-plugin:acceptPlugin
