--- build.gradle.kts
@@ -1,76 +1,0 @@
-import com.github.benmanes.gradle.versions.updates.DependencyUpdatesTask
-import io.gitlab.arturbosch.detekt.Detekt
-
-plugins {
-    alias(libs.plugins.kotlin) apply false
-    alias(libs.plugins.detekt)
-    alias(libs.plugins.ktlint)
-    alias(libs.plugins.versionCheck)
-    idea
-}
-
-subprojects {
-    apply {
-        plugin(rootProject.libs.plugins.detekt.get().pluginId)
-        plugin(rootProject.libs.plugins.ktlint.get().pluginId)
-        plugin("org.gradle.idea")
-    }
-
-    ktlint {
-        debug.set(false)
-        verbose.set(true)
-        android.set(false)
-        outputToConsole.set(true)
-        ignoreFailures.set(false)
-        enableExperimentalRules.set(true)
-        filter {
-            exclude("**/generated/**")
-            include("**/kotlin/**")
-        }
-    }
-
-    detekt {
-        config.setFrom(rootProject.files("config/detekt/detekt.yml"))
-    }
-
-    idea {
-        module {
-            isDownloadJavadoc = true
-            isDownloadSources = true
-        }
-    }
-}
-
-tasks.withType<Detekt>().configureEach {
-    reports {
-        html.required.set(true)
-        html.outputLocation.set(file("build/reports/detekt.html"))
-    }
-}
-
-tasks.withType<DependencyUpdatesTask> {
-    rejectVersionIf {
-        candidate.version.isNonStable()
-    }
-}
-
-fun String.isNonStable() = "^[0-9,.v-]+(-r)?$".toRegex().matches(this).not()
-
-tasks.register("reformatAll") {
-    description = "Reformat all the Kotlin Code"
-
-    dependsOn(":ktlintFormat")
-    dependsOn(gradle.includedBuild("kover-badge-plugin").task("ktlintFormat"))
-}
-
-tasks.register("preMerge") {
-    description = "Runs all the tests/verification tasks on both top level and included build."
-
-    println(":check")
-    dependsOn(gradle.includedBuild("kover-badge-plugin").task(":check"))
-    dependsOn(gradle.includedBuild("kover-badge-plugin").task(":validatePlugins"))
-}
-
-tasks.wrapper {
-    distributionType = Wrapper.DistributionType.ALL
-}
