+++ buildSrc/settings.gradle.kts
@@ -1,0 +1,9 @@
+rootProject.name = "buildSrc"
+
+dependencyResolutionManagement {
+    versionCatalogs {
+        create("libs") {
+            from(files("../gradle/libs.versions.toml"))
+        }
+    }
+}
