+++ plugins/build.gradle.kts
@@ -1,0 +1,8 @@
+plugins {
+    kotlin("jvm")
+}
+
+kotlin {
+    jvmToolchain(libs.versions.java.get().toInt())
+}
+
