+++ plugins/common/build.gradle.kts
@@ -1,0 +1,23 @@
+plugins {
+    kotlin("jvm")
+}
+
+group = "io.flax"
+
+kotlin {
+    jvmToolchain(libs.versions.java.get().toInt())
+}
+
+tasks.withType<org.jetbrains.kotlin.gradle.tasks.KotlinCompile>().configureEach {
+    kotlinOptions {
+        freeCompilerArgs = freeCompilerArgs + "-Xcontext-receivers"
+    }
+}
+
+repositories {
+    mavenCentral()
+}
+
+dependencies {
+    implementation(gradleApi())
+}
