+++ plugins/kover-badge-plugin/gradle.properties
