+++ plugins/kover-badge-plugin/src/functionalTest/resources/test-project-files/README
