+++ plugins/kover-badge-plugin/src/functionalTest/resources/test-project-files/build-gradle
