+++ plugins/kover-badge-plugin/src/test/kotlin/io/flax/kover/KoverBadgePluginTest.kt
