--- renovate.json
@@ -1,6 +1,0 @@
-{
-  "extends": [
-    "config:base",
-    ":automergeMinor"
-  ]
-}
