--- test-project/TEST_README.md
@@ -1,2 +1,0 @@
-# Test Readme
-<a href="file:/Users/ido/IdeaProjects/flax-gradle-plugins/test-project/build/reports/kover/html/index.html">![koverage](https://img.shields.io/badge/100.0-green?logo=kotlin&label=koverage&style=plastic)</a>
