+++ test-projects/kover-badge-plugin/TEST_README.md
@@ -1,0 +1,1 @@
+<a href="file:/Users/ido/IdeaProjects/flax-gradle-plugins-copy/test-projects/kover-badge-plugin/build/reports/kover/html/index.html">![koverage](https://img.shields.io/badge/100.0-green?logo=kotlin&label=koverage&style=plastic)</a>
