+++ test-projects/kover-badge-plugin/src/main/kotlin/io/flax/kover/Application.kt
