+++ test-projects/kover-badge-plugin/src/test/kotlin/io/flax/kover/ApplicationTest.kt
